requirements.txt
tinkoff-investments
pandas
numpy
ta
python-dotenv
matplotlib